refactor: re-organize dotnet reference generation
