+++ src/Docfx.Dotnet/ManagedReference/ApiReferenceViewModel.cs
