+++ src/Docfx.Dotnet/ManagedReference/ExtractMetadataConfig.cs
