+++ src/Docfx.Dotnet/ManagedReference/ExtractMetadataException.cs
