+++ src/Docfx.Dotnet/ManagedReference/MetadataItem.cs
