+++ src/Docfx.Dotnet/ManagedReference/MetadataModel.cs
