+++ src/Docfx.Dotnet/ManagedReference/Models/AdditionalNotes.cs
