+++ src/Docfx.Dotnet/ManagedReference/Models/ExceptionInfo.cs
