+++ src/Docfx.Dotnet/ManagedReference/Models/ItemViewModel.cs
