+++ src/Docfx.Dotnet/ManagedReference/Models/LinkInfo.cs
