+++ src/Docfx.Dotnet/ManagedReference/Models/SyntaxDetailViewModel.cs
