+++ src/Docfx.Dotnet/ManagedReference/ReferenceItem.cs
