+++ src/Docfx.Dotnet/ManagedReference/Resolvers/BuildToc.cs
