+++ src/Docfx.Dotnet/ManagedReference/Resolvers/IResolverPipeline.cs
