+++ src/Docfx.Dotnet/ManagedReference/Resolvers/LayoutCheckAndCleanup.cs
