+++ src/Docfx.Dotnet/ManagedReference/Resolvers/NormalizeSyntax.cs
