+++ src/Docfx.Dotnet/ManagedReference/Resolvers/ResolveReference.cs
