+++ src/Docfx.Dotnet/ManagedReference/Resolvers/ResolverContext.cs
