+++ src/Docfx.Dotnet/ManagedReference/Resolvers/SetDerivedClass.cs
