+++ src/Docfx.Dotnet/ManagedReference/Resolvers/SetParent.cs
