+++ src/Docfx.Dotnet/ManagedReference/Resolvers/YamlMetadataResolver.cs
