+++ src/Docfx.Dotnet/ManagedReference/SyntaxDetail.cs
