+++ src/Docfx.Dotnet/ManagedReference/Visitors/SpecIdCoreVisitor.cs
