+++ src/Docfx.Dotnet/ManagedReference/Visitors/SpecIdHelper.cs
