+++ src/Docfx.Dotnet/ManagedReference/Visitors/SymbolVisitorAdapter.cs
