+++ src/Docfx.Dotnet/ManagedReference/Visitors/VisitorHelper.cs
