+++ src/Docfx.Dotnet/ManagedReference/Visitors/YamlModelGenerator.cs
